Starting refactoring of Administrator.java + Orders equals() fix

+++ src/restaurant/util/Constants.java
@@ -5,6 +5,6 @@ public class Constants {
     public static final Object[] DISHES_TABLE_HEADER = {
             "Name", "Description", "Cooking time", "Price"
     };
-    public static final long UPDATE_DELAY = 5000; // 5 sec
+    public static final int UPDATE_DELAY = 5000; // 5 sec
     public static final int MENU_VISIBLE_FIELDS = 4;
 }
